feat: segragate MarkdownEditor
